SourceFileUtils is singleton, creation window still needs to read files
Temporarily changed SourceFileUtils as a singleton object that has to be instanced once in order to pass "project" and "windowtoobox" references
that can be used later and there is no need to constantly pass project reference in almost every method. 
A Runtime Exception has been written for that case, very basic. Object Creation Controller will take over most logic from GUI in the next iteration.
ObjectWindowFactory instantiates SourceFileUtils.
Changed version in Gradle to reflect coming to the end of the iteration (0.1.9)