Recursive Class search and Loading Complete
Classes are loaded appropriately and are found in the java compile folder regardless of place (however there is is a possibility of it loading a wrong folder if there are multiple packages with build folder, additional check for *.java and *.class files is needed). SourceFileUtils is a singleton object now. Removed passing toolWindow and project reference to the ObjectDisplayWindow  as SoutceFileUtils now stores it.