External library button has been disabled.
External jar loading now happens as a part of regular class loading method, those external libraries are not part of the object workbench and do not show on the object creation, nor graph. There might be an additional feature considered where external libraries are loaded so they can be seen.