Object instantiation is now working in 50%
It can pull information from text boxes, however they are always String instances so additional work is needed there. Also if a constructor take an object instance, there is a need for a drop-down menu listing possible objects to put into it. This might be problematic if it requires something from external library that is not part of either standard library or project's scope.